Atualizado codigo para processamento. Criada base com menor numero de fragmentos

+++ montagemFragmentos/include/fragment.h
@@ -11,11 +11,11 @@
 typedef unsigned int uint;
 
 typedef struct t_fragment {
-	int id;
+	long id;
 	char* sequence;
 	struct t_fragment* next;
 } Fragment;
 
-Fragment* buildFragment(int id, char* sequence, uint n);
+Fragment* buildFragment(long id, char* sequence, uint n);
 
 #endif /* FRAGMENT_H_ */
